segundo commit

+++ src/main/java/br/com/globalhitss/avaliacao/repositories/VeiculoRepository.java
@@ -2,7 +2,7 @@ package br.com.globalhitss.avaliacao.repositories;
 
 import br.com.globalhitss.avaliacao.model.Veiculo;
 import org.springframework.data.jpa.repository.JpaRepository;
-import java.math.BigInteger;
 
-public interface VeiculoRepository extends JpaRepository<Veiculo, BigInteger> {
+
+public interface VeiculoRepository extends JpaRepository<Veiculo, Long> {
 }
